Always create containers.
Consensus from user testing at Emerald Sprint is that this option
can be confusing to a user, and creating a container is usually
the safer bet and has minimal extra overhead. We're adding UI
for configuring which types can be added, so it will still be
possible to disallow adding any contained types.