Update docu for creating dexterity behaviors
Adjusted imports etc. so they work for Plone 5.1.2
Note to reviewer: we might want to use a more recent package (not something on svn.plone.org) as an example for dexterity docu. If you could point me to a good example I'd be happy to adjust the docu to this.